Factor out title extraction
