+++ packages/react-server-website/components/doc-contents.js
@@ -1,16 +1,14 @@
 import React from "react";
 import {join} from "path";
+import PageNameMixin from "../lib/page-name-mixin";
 import {
 	Link,
 	getCurrentRequestContext,
-	RequestLocalStorage,
 } from "react-server";
 
 import SvgDropdown from './assets/SvgDropdown';
 
 import './doc-contents.less'
-
-const RLS = RequestLocalStorage.getNamespace();
 
 const ContentsSection = ({name, pages}) => (
 	<div className='contentsSection'>
@@ -39,24 +37,6 @@ export default class DocContents extends React.Component {
 		};
 	}
 
-	static setResponse(res) {
-		const path = getCurrentRequestContext().getCurrentPath().replace("/docs/", "");
-
-		// This is all we care about stashing away.  We'll _also_ receive the
-		// response as a prop on our instance, later.  We just need the active
-		// page _before_ our element is created (for the page title, etc)
-		RLS().activePageName = (res.contents.reduce((page, section) => (
-			page || section.pages.find(page => page.path === path)
-		), null) || {name: "React Server Documentation"}).name;
-
-		// Pass it along.
-		return res;
-	}
-
-	static activePageName() {
-		return RLS().activePageName;
-	}
-
 	componentDidMount() {
 		getCurrentRequestContext().navigator.on( "navigateStart", this.closeMenu.bind(this) );
 	}
@@ -81,3 +61,8 @@ export default class DocContents extends React.Component {
 		this.setState( {menuOpen: false} );
 	}
 }
+
+PageNameMixin(DocContents, {
+	prefix: "/docs/",
+	defaultName: "React Server Documentation",
+});
