+++ packages/react-server-website/components/source-contents.js
@@ -1,14 +1,12 @@
 import React from "react";
 import {join} from "path";
+import PageNameMixin from "../lib/page-name-mixin";
 import {
 	Link,
 	getCurrentRequestContext,
-	RequestLocalStorage,
 } from "react-server";
 
 import './doc-contents.less'
-
-const RLS = RequestLocalStorage.getNamespace();
 
 const SourceContentsSection = ({name, pages}) => <div className='contentsSection'>
 	<h3>{name}</h3>
@@ -28,28 +26,14 @@ const ContentsLinkWithMungedPath = (name, path) => <li {...classIfActive(path)}>
 </li>
 
 export default class SourceContents extends React.Component {
-
-	static setResponse(res) {
-		const path = getCurrentRequestContext().getCurrentPath().replace("/source/", "");
-
-		// This is all we care about stashing away.  We'll _also_ receive the
-		// response as a prop on our instance, later.  We just need the active
-		// page _before_ our element is created (for the page title, etc)
-		RLS().activePageName = (res.contents.reduce((page, section) => (
-			page || section.pages.find(page => page.path === path)
-		), null) || {name: "React Server Documentation"}).name;
-
-		// Pass it along.
-		return res;
-	}
-
-	static activePageName() {
-		return RLS().activePageName;
-	}
-
 	render() {
 		return <div className='DocContents'>{
 			this.props.contents.map(SourceContentsSection)
 		}</div>
 	}
 }
+
+PageNameMixin(SourceContents, {
+	prefix: "/source/",
+	defaultName: "React Server Website Source",
+});
